fix(ci): make the lint gate actually pass — it was verified with a broken check
The `lint` job added in the previous commit failed on push. The gate was fine;
the way I checked it locally was not.

I verified ruff with `ruff check . -q | grep '^Found'`. `-q` SUPPRESSES that
summary line, so the grep could never match and the check reported "clean"
whether or not there were errors. A check that cannot fail is not a check —
the same defect class the SQL guardrail work was about, in the verification of
that very work. Now checked by exit code.

What was actually broken: `I001` import ordering, because inserting
`from typing import Any` put it above the stdlib block; plus a few
`RUF046`/`F841`/`RUF059`/`SIM105` that were never fixed because the broken
check said they were.

Test jobs were green throughout; only lint was red.

+++ src/can_telemetry_forge/config.py
@@ -245,7 +245,7 @@ class ForgeConfig:
 
     def n_steps(self) -> int:
         """Number of timestamps per unit over the window."""
-        return int(round(self.days * 24.0 / self.step_hours()))
+        return round(self.days * 24.0 / self.step_hours())
 
     def rng(self) -> np.random.Generator:
         """The master seeded generator (ADR-005). Child rngs spawn from this."""
